Change permutation repr to u16

+++ tsp_samplers/src/algorithms.rs
@@ -1,8 +1,8 @@
 use rustc_hash::FxHashMap;
 
-pub type HillclimbFunction = fn(&Vec<usize>, &Vec<Vec<i32>>, bool) -> (Vec<usize>, i32);
-pub type NodeMap = FxHashMap<Vec<usize>, (u32, i32)>;
-pub type EdgeMap = FxHashMap<(u32, u32), i32>;
+pub type HillclimbFunction = fn(&Vec<u16>, &Vec<Vec<i32>>, bool) -> (Vec<u16>, i32);
+pub type NodeMap = FxHashMap<Vec<u16>, (u16, i32)>;
+pub type EdgeMap = FxHashMap<(u16, u16), i32>;
 
 pub mod snowball_sampler;
 pub mod exhaustive_sampler;
